requirements.txt
fastapi
uvicorn[standard]
celery[redis]
redis
pygithub
python-dotenv
pytest
httpx
ollama